requirements.txt.txt
google-generativeai
streamlit
os